Doc_react (1)/Doc_react/django_rest_api/backend/api/serializers.py
from rest_framework import serializers
from .models import UploadFiles

class UploadFileSerializer(serializers.ModelSerializer):
    class Meta:
        model = UploadFiles
        fields = ['uploadedfile']