Doc_react (1)/Doc_react/django_rest_api/backend/requirements.txt
django>=4.0.0,<4.1.0
djangorestframework
pyyaml
requests
django-cors-headers
pymupdf
opencv-python
beautifulSoup4
python-docx
layoutparser
layoutparser[ocr]
pytest
pytesseract
pycocotools
paddlepaddle